init onEnter prop for redirection:
the passed in function to onEnter is called when a route is about to be
entered. In this case, when user access to detail route, the assignShow
function is activated. This function checks if requested url id matches
any of instances of data set. If none of them matches, it redirects to
main(/) page. If any data was found, the data is passed into params.