update Header and Search:
For component method's reusability, handleSearchTermEvent in the Search
component is changed to handleSearchTermChange. Instead of picking the
value directly from the child, it just takes input and set the value to
searchTerm.

Now the child component(header) takes value with the method
handleSearchTermEvent which gets value and pass the value to the passed
down parent's method.

This kind of approach is crutial for code or component reusability.